+++ build-all-branches.bash
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 test -f "$HOME/.jdk8" && . $HOME/.jdk8
 
-for v in "183.4139.22" "2018.2.3" "2018.1.6" "2017.3.5" "2017.2.6" "2017.1.6" "2016.3.8" "2016.2.5" "2016.1.4"; do
+for v in "191.4212.41" "183.4139.22" "2018.2.3" "2018.1.6" "2017.3.5" "2017.2.6" "2017.1.6" "2016.3.8" "2016.2.5" "2016.1.4"; do
     echo "## Building with version $v..."
     _JAVA_OPTIONS="" JAVA_OPTS="" gradle -Dbash.skipUrls="true" -PideaVersion="$v" -PideaBranch="" clean build