+++ test/com/ansorgit/plugins/bash/editor/inspections/inspections/MissingIncludeFileInspectionTest.java
@@ -41,7 +41,7 @@ public class MissingIncludeFileInspectionTest extends AbstractInspectionTestCase
             return;
         }
 
-        BashProjectSettings settings = BashProjectSettings.storedSettings(myProject);
+        BashProjectSettings settings = BashProjectSettings.storedSettings(getProject());
         boolean old = settings.isValidateWithCurrentEnv();
         try {
             settings.setValidateWithCurrentEnv(false);
@@ -59,5 +59,10 @@ public class MissingIncludeFileInspectionTest extends AbstractInspectionTestCase
     @Test
     public void testIncludeDirectory() {
         doTest("missingIncludeFileInspection/includeDirectory", new MissingIncludeFileInspection());
+    }
+
+    @Test
+    public void testInvalidIncludePath() {
+        doTest("missingIncludeFileInspection/invalidIncludePath", new MissingIncludeFileInspection());
     }
 }